Test accuracy issues solved.
There was a issue with the test accuracy of the simple classifier model and complex classifier model. Previously, It both of the model was giving the similar accuracy as the pre-trained model.  I detect the issues and changed the variable names to meaningful variable and now it shows the genuine accuracy individually.